requirements.txt
alabaster==0.7.13
appdirs==1.4.4
attrs==23.1.0
Babel==2.12.1
bleach==6.0.0
cattrs==23.1.2
certifi==2023.5.7
charset-normalizer==3.2.0
colorama==0.4.6
docutils==0.18.1
esbonio==0.16.1
idna==3.4
imagesize==1.4.1
importlib-metadata==6.8.0
iniconfig==2.0.0
jaraco.classes==3.3.0
Jinja2==3.1.2
keyring==24.2.0
lsprotocol==2023.0.0a2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
more-itertools==9.1.0
packaging==23.1
pkginfo==1.9.6
pluggy==1.2.0
pygls==1.0.2
Pygments==2.15.1
pyspellchecker==0.7.2
pytest==7.4.0
python-dotenv==1.0.0
python-ndev-blizzardapi==0.1.0
pywin32-ctypes==0.2.2
readme-renderer==40.0
requests==2.31.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.4.2
six==1.16.0
snowballstemmer==2.2.0
Sphinx==6.2.1
sphinx-rtd-theme==1.2.2
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
twine==4.0.2
typeguard==3.0.2
urllib3==2.0.3
webencodings==0.5.1
zipp==3.16.2
